% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAxisLabel.R
\name{addAxisLabel}
\alias{addAxisLabel}
\title{addAxisLabel()}
\usage{
addAxisLabel(myIO, xAxis = NULL, yAxis = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a string label to be printed along the x axis}

\item{yAxis}{a string label to be printedalong the y axis}
}
\value{
the same myIO object
}
\description{
Suppresses axes from printing
}
