% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIoLayer.R
\name{addIoLayer}
\alias{addIoLayer}
\title{Add a Layer to a myIO Chart}
\usage{
addIoLayer(
  myIO,
  type,
  color = NULL,
  label,
  data = NULL,
  mapping,
  transform = "identity",
  options = list(barSize = "large", toolTipOptions = list(suppressY = FALSE))
)
}
\arguments{
\item{myIO}{an htmlwidget object created by the \code{myIO()} function}

\item{type}{chart type}

\item{color}{optional CSS color string or vector for grouped layers}

\item{label}{unique layer label}

\item{data}{data frame backing the layer}

\item{mapping}{named aesthetic mapping list}

\item{transform}{transform name applied before serialization}

\item{options}{layer options passed through to the widget config}
}
\value{
A modified \code{myIO} htmlwidget object with the new layer appended
  to the configuration.
}
\description{
Adds individual layer to a myIO widget
}
\examples{
myIO(data = mtcars) |>
  addIoLayer(
    type = "point", label = "points",
    mapping = list(x_var = "wt", y_var = "mpg")
  )

}
