% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_boxplot.R
\name{composite_boxplot}
\alias{composite_boxplot}
\title{Boxplot composite expansion}
\usage{
composite_boxplot(data, mapping, label, color, options)
}
\description{
Boxplot composite expansion
}
\keyword{internal}
