% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_comparison.R
\name{composite_comparison}
\alias{composite_comparison}
\title{Comparison composite expansion}
\usage{
composite_comparison(data, mapping, label, color, options)
}
\description{
Comparison composite expansion
}
\keyword{internal}
