% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_fan.R
\name{composite_fan}
\alias{composite_fan}
\title{Fan chart composite expansion}
\usage{
composite_fan(data, mapping, label, color, options)
}
\description{
Fan chart composite expansion
}
\keyword{internal}
