% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_regression.R
\name{composite_regression}
\alias{composite_regression}
\title{Regression composite}
\usage{
composite_regression(data, mapping, label, color, options)
}
\description{
Expands into scatter + trend line + CI band + R² annotation sublayers.
}
\keyword{internal}
