% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_ridgeline.R
\name{composite_ridgeline}
\alias{composite_ridgeline}
\title{Ridgeline composite expansion}
\usage{
composite_ridgeline(data, mapping, label, color, options)
}
\description{
Ridgeline composite expansion
}
\keyword{internal}
