% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_violin.R
\name{composite_violin}
\alias{composite_violin}
\title{Violin composite expansion}
\usage{
composite_violin(data, mapping, label, color, options)
}
\description{
Violin composite expansion
}
\keyword{internal}
