% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCategoricalAxis.R
\name{defineCategoricalAxis}
\alias{defineCategoricalAxis}
\title{Define Categorical Axis}
\usage{
defineCategoricalAxis(myIO, xAxis = TRUE, yAxis = FALSE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a logical argument (TRUE) for defining the x axis as categorical}

\item{yAxis}{a logical argument (TRUE) for defining the y axis as categorical}
}
\value{
A modified \code{myIO} htmlwidget object with categorical axis
  configured.
}
\description{
Function to define the x variable as categorical
}
\examples{
# Define x axis as categorical
myIO() |> defineCategoricalAxis(xAxis = TRUE)

# Define both axes as categorical
myIO() |> defineCategoricalAxis(xAxis = TRUE, yAxis = TRUE)

}
