% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragPoints.R
\name{dragPoints}
\alias{dragPoints}
\title{Enable Draggable Points}
\usage{
dragPoints(myIO, dragPoints = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{dragPoints}{a logical argument (TRUE) for creating draggable points}
}
\value{
A modified \code{myIO} htmlwidget object with drag interaction
  enabled.
}
\description{
Function to make points draggable
}
\examples{
# Enable draggable points
myIO() |>
  addIoLayer(
    type = "point", color = "red", label = "pts",
    data = mtcars, mapping = list(x_var = "wt", y_var = "mpg")
  ) |>
  dragPoints()

}
