% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_expand.R
\name{expandComposite}
\alias{expandComposite}
\title{Expand a composite layer into sub-layers}
\usage{
expandComposite(type, label, data, mapping, options, color, existing_layers)
}
\description{
Expand a composite layer into sub-layers
}
\keyword{internal}
