% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipAxis.R
\name{flipAxis}
\alias{flipAxis}
\title{Flip Chart Axes}
\usage{
flipAxis(myIO, flipAxis = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{flipAxis}{a logical argument (TRUE) for flipping the x and y axes}
}
\value{
A modified \code{myIO} htmlwidget object with axes flipped.
}
\description{
Function to flip the x and y axes
}
\examples{
# Flip the axes for a horizontal bar chart
myIO() |>
  addIoLayer(
    type = "bar", color = "steelblue", label = "bars",
    data = mtcars, mapping = list(x_var = "wt", y_var = "mpg")
  ) |>
  flipAxis()

}
