% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIO.R
\name{myIO-shiny}
\alias{myIO-shiny}
\alias{myIOOutput}
\alias{renderMyIO}
\title{Shiny Bindings for myIO}
\usage{
myIOOutput(outputId, width = "100\%", height = "400px")

renderMyIO(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit or a number.}

\item{expr}{An expression that generates a myIO}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\value{
\code{myIOOutput} returns a Shiny UI element for placement in a UI
  definition. \code{renderMyIO} returns a Shiny render function for use in a
  server definition.
}
\description{
Shiny Bindings for myIO
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(myIOOutput("chart"))
  server <- function(input, output) {
    output$chart <- renderMyIO({
      myIO(data = mtcars) |>
        addIoLayer(type = "point", label = "scatter",
          mapping = list(x_var = "wt", y_var = "mpg"))
    })
  }
  shinyApp(ui, server)
}

}
