% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAxisLimits.R
\name{setAxisLimits}
\alias{setAxisLimits}
\title{Set Axis Limits}
\usage{
setAxisLimits(
  myIO,
  xlim = list(min = NULL, max = NULL),
  ylim = list(min = NULL, max = NULL)
)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xlim}{a list of min and max values}

\item{ylim}{a list of min and max values}
}
\value{
A modified \code{myIO} htmlwidget object with updated axis limits.
}
\description{
Sets the minimum and maximum values for chart axes
}
\examples{
# Set x axis limits
myIO() |> setAxisLimits(xlim = list(min = 0, max = 100))

# Set both axis limits
myIO() |> setAxisLimits(
  xlim = list(min = 0, max = 50),
  ylim = list(min = -10, max = 200)
)

}
