% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBrush.R
\name{setBrush}
\alias{setBrush}
\title{Enable Brush Selection}
\usage{
setBrush(myIO, direction = "xy", on_select = "highlight")
}
\arguments{
\item{myIO}{an htmlwidget object created by the \code{myIO()} function}

\item{direction}{brush direction: \code{"xy"} (default), \code{"x"},
or \code{"y"}}

\item{on_select}{behavior in static mode: \code{"highlight"} (default)
or \code{"export"} (scopes CSV download to selected points)}
}
\value{
A modified \code{myIO} htmlwidget object with brush interaction
  enabled.
}
\description{
Enables rectangle brush selection on the chart. When a user drags to
select a region, the selected data points are available as a reactive
input in Shiny or exportable in static HTML.
}
\examples{
myIO(data = mtcars) |>
  addIoLayer(
    type = "point", label = "pts",
    mapping = list(x_var = "wt", y_var = "mpg")
  ) |>
  setBrush()

}
