% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLayerOpacity.R
\name{setLayerOpacity}
\alias{setLayerOpacity}
\title{Set Layer Opacity}
\usage{
setLayerOpacity(myIO, label, opacity)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{label}{character. The layer label to target.}

\item{opacity}{numeric. Opacity value between 0 (transparent) and 1 (opaque).}
}
\value{
A modified \code{myIO} htmlwidget object.
}
\description{
Set Layer Opacity
}
\examples{
myIO(iris) |>
  addIoLayer("point", label = "pts",
             mapping = list(x_var = "Sepal.Length", y_var = "Sepal.Width")) |>
  setLayerOpacity("pts", 0.5)

}
