% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReferenceLines.R
\name{setReferenceLines}
\alias{setReferenceLines}
\title{Set Reference Lines}
\usage{
setReferenceLines(myIO, xRef = 0, yRef = 0)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xRef}{a list of the reference line value of x}

\item{yRef}{a list of the reference line value of y}
}
\value{
A modified \code{myIO} htmlwidget object with reference lines added.
}
\description{
Sets x and y reference lines
}
\examples{
# Add reference lines at x=5 and y=20
myIO() |> setReferenceLines(xRef = 5, yRef = 20)

}
