% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTitle.R
\name{setTitle}
\alias{setTitle}
\title{Set Chart Title}
\usage{
setTitle(myIO, title = NULL)
}
\arguments{
\item{myIO}{A myIO widget object.}

\item{title}{Character title or NULL to remove the title.}
}
\value{
A modified \code{myIO} htmlwidget object.
}
\description{
Sets the title rendered inside the myIO widget SVG.
}
\examples{
myIO() |>
  setTitle("Miles per gallon") |>
  addIoLayer("point", label = "cars",
             data = mtcars,
             mapping = list(x_var = "wt", y_var = "mpg"))

}
