% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setToolTipOptions.R
\name{setToolTipOptions}
\alias{setToolTipOptions}
\title{Set Tooltip Options}
\usage{
setToolTipOptions(myIO, suppressY = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{suppressY}{a boolean}
}
\value{
A modified \code{myIO} htmlwidget object with updated tooltip
  options.
}
\description{
Generic function for setting tool tip options for a chart
}
\examples{
# Suppress the y value in tooltips
myIO() |> setToolTipOptions(suppressY = TRUE)

}
