% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressAxis.R
\name{suppressAxis}
\alias{suppressAxis}
\title{Suppress Axis Display}
\usage{
suppressAxis(myIO, xAxis = NULL, yAxis = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a logical operator defining whether the x axis should be printed or not}

\item{yAxis}{a logical operator defining whether the y axis should be printed or not}
}
\value{
A modified \code{myIO} htmlwidget object with axis display
  suppressed.
}
\description{
Suppresses axes from printing
}
\examples{
# Suppress both axes
myIO() |> suppressAxis(xAxis = TRUE, yAxis = TRUE)

# Suppress only the x axis
myIO() |> suppressAxis(xAxis = TRUE)

}
