% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressLegend.R
\name{suppressLegend}
\alias{suppressLegend}
\title{Suppress Legend Display}
\usage{
suppressLegend(myIO, suppressLegend = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{suppressLegend}{a logical operator defining whether the legend should be printed or not}
}
\value{
A modified \code{myIO} htmlwidget object with legend display
  suppressed.
}
\description{
Suppresses legend from printing
}
\examples{
# Hide the legend
myIO() |> suppressLegend()

}
