% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_ci.R
\name{transform_ci}
\alias{transform_ci}
\title{Confidence / prediction interval transform}
\usage{
transform_ci(data, mapping, options = list())
}
\description{
Confidence / prediction interval transform
}
\keyword{internal}
