% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_cumulative.R
\name{transform_cumulative}
\alias{transform_cumulative}
\title{Cumulative transform}
\usage{
transform_cumulative(data, mapping, options = list())
}
\description{
Cumulative transform
}
\keyword{internal}
