% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_mean_ci.R
\name{transform_mean_ci}
\alias{transform_mean_ci}
\title{Group mean with confidence interval transform}
\usage{
transform_mean_ci(data, mapping, options = list())
}
\description{
Group mean with confidence interval transform
}
\keyword{internal}
