% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pairwise_test.R
\name{transform_pairwise_test}
\alias{transform_pairwise_test}
\title{Pairwise hypothesis test transform}
\usage{
transform_pairwise_test(data, mapping, options = list())
}
\description{
Pairwise hypothesis test transform
}
\keyword{internal}
