% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_polynomial.R
\name{transform_polynomial}
\alias{transform_polynomial}
\title{Polynomial regression transform}
\usage{
transform_polynomial(data, mapping, options = list())
}
\description{
Polynomial regression transform
}
\keyword{internal}
