% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quantile_dots.R
\name{transform_quantile_dots}
\alias{transform_quantile_dots}
\title{Quantile dots transform}
\usage{
transform_quantile_dots(data, mapping, options = list())
}
\description{
Quantile dots transform
}
\keyword{internal}
