% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quantiles.R
\name{transform_quantiles}
\alias{transform_quantiles}
\title{Quantiles transform}
\usage{
transform_quantiles(data, mapping, options = list())
}
\description{
Quantiles transform
}
\keyword{internal}
