% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_residuals.R
\name{transform_residuals}
\alias{transform_residuals}
\title{Regression residuals transform}
\usage{
transform_residuals(data, mapping, options = list())
}
\description{
Regression residuals transform
}
\keyword{internal}
