% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_smooth.R
\name{transform_smooth}
\alias{transform_smooth}
\title{Moving average / exponential smoothing transform}
\usage{
transform_smooth(data, mapping, options = list())
}
\description{
Moving average / exponential smoothing transform
}
\keyword{internal}
