% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_summary.R
\name{transform_summary}
\alias{transform_summary}
\title{Summary statistics transform}
\usage{
transform_summary(data, mapping, options = list())
}
\description{
Summary statistics transform
}
\keyword{internal}
