% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIoLayer.R
\name{addIoLayer}
\alias{addIoLayer}
\title{addLayer()}
\usage{
addIoLayer(myIO, type, color = NULL, label, data = NULL, mapping,
  options = list(barSize = "large", toolTipOptions = list(suppressY =
  FALSE)))
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{type}{a quoted string defining the type of layer: 'bar', 'line', 'point', 'hexbin', 'treemap'}

\item{color}{a quoted string defining the layer's color as an CSS recognized color}

\item{label}{a unique quoted label/id string for the plot}

\item{data}{(optional) R dataframe/tibble}

\item{mapping}{a list object defining variables for the layer: layer(x_var = "xVar", y_var = "yVar")}

\item{options}{a list oboject defining options for that layer}
}
\value{
the same myIO object with the new layer attached to the message for the browser (myIO$x$layers)
}
\description{
Adds individual layer to a myIO widget
}
