% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_duckdb_wasm.R
\name{clear_duckdb_wasm_cache}
\alias{clear_duckdb_wasm_cache}
\title{Remove DuckDB-WASM cache entries}
\usage{
clear_duckdb_wasm_cache(version = NULL)
}
\arguments{
\item{version}{Character scalar naming a specific version to remove. If
NULL, removes all cached versions.}
}
\value{
Number of removed entries, invisibly.
}
\description{
Remove DuckDB-WASM cache entries
}
