% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCategoricalAxis.R
\name{defineCategoricalAxis}
\alias{defineCategoricalAxis}
\title{defineCategoricalAxis()}
\usage{
defineCategoricalAxis(myIO, xAxis = TRUE, yAxis = FALSE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a logical argument (TRUE) for defining the x axis as categrocial}

\item{yAxis}{a logical argument (TRUE) for defining the y axis as categrocial}
}
\value{
the same myIO object
}
\description{
Function to define the x variable as categorical
}
