% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCategoricalAxis.R
\name{defineCategoricalAxis}
\alias{defineCategoricalAxis}
\title{defineCategoricalAxis()}
\usage{
defineCategoricalAxis(myIO, categoricalScale = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{categoricalScale}{a logical argument (TRUE) for defining the x as categrocial}
}
\value{
the same myIO object
}
\description{
Function to define the x variable as categorical
}
