% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragPoints.R
\name{dragPoints}
\alias{dragPoints}
\title{dragPoints()}
\usage{
dragPoints(myIO, dragPoints = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{dragPoints}{a logical argument (TRUE) for creating draggable points}
}
\value{
the same myIO object
}
\description{
Function to make points draggable
}
