% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_duckdb_wasm.R
\name{duckdb_wasm_status}
\alias{duckdb_wasm_status}
\title{DuckDB-WASM cache status}
\usage{
duckdb_wasm_status()
}
\value{
A list with class `myIO_duckdb_wasm_status` and fields
  `installed` (logical), `version` (chr or NA), `cache_dir` (chr),
  `size_bytes` (numeric).
}
\description{
DuckDB-WASM cache status
}
\examples{
duckdb_wasm_status()
}
