% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipAxis.R
\name{flipAxis}
\alias{flipAxis}
\title{flipAxis()}
\usage{
flipAxis(myIO, flipAxis = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{flipAxis}{a logical argument (TRUE) for flipping the x and y axes}
}
\value{
the same myIO object
}
\description{
Function to flip the x and y axes
}
