% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIO-package.R
\docType{package}
\name{myIO-package}
\alias{myIO-package}
\title{myIO: Interactive Data Visualizations Using d3.js}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create interactive 'd3.js' visualizations from R with built-in statistical transforms. Computes confidence intervals, regression fits, LOESS smoothing, moving averages, and error bars in R and renders them as composable chart layers via 'htmlwidgets'. Supports 20 chart types including boxplots, violin plots, Q-Q diagnostic plots, and group comparisons with pairwise significance testing. Works in 'RStudio', 'Shiny', and 'R Markdown'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mortonanalytics.github.io/myIO/}
  \item \url{https://github.com/mortonanalytics/myIO}
  \item Report bugs at \url{https://github.com/mortonanalytics/myIO/issues}
}

}
\author{
\strong{Maintainer}: Ryan Morton \email{morton@myma.us} [copyright holder]

}
\keyword{internal}
