% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIO.R
\name{myIO-shiny}
\alias{myIO-shiny}
\alias{myIOOutput}
\alias{renderMyIO}
\title{Shiny bindings for myIO}
\usage{
myIOOutput(outputId, width = "100\%", height = "400px")

renderMyIO(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit or a number.}

\item{expr}{An expression that generates a myIO}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\description{
Shiny bindings for myIO
}
