% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIO.R
\name{myIO}
\alias{myIO}
\title{Create a myIO Chart Widget}
\usage{
myIO(
  data = NULL,
  width = "100\%",
  height = "400px",
  elementId = NULL,
  sparkline = FALSE
)
}
\arguments{
\item{data}{an optional point of entry for the data frame or vector}

\item{width}{a string of either pixel width or a percentage width}

\item{height}{a string of pixel height}

\item{elementId}{a unique id for the htmlwidget object}

\item{sparkline}{Logical. If TRUE, renders a compact sparkline suitable for
embedding in table cells. Strips axes, legend, and interactions.
Only "line", "bar", and "area" layer types are supported. Default FALSE.}
}
\value{
An htmlwidget object of class \code{myIO}.
}
\description{
Create an interactive D3.js chart widget
}
\examples{
myIO(data = mtcars) |>
  setMargin(top = 40, bottom = 80, left = 60, right = 10)

}
