% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIO.R
\name{myIO}
\alias{myIO}
\title{myIO}
\usage{
myIO(data = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{an optional point of entry for the data frame or vector}

\item{width}{a string of either pixel width or a percentage width}

\item{height}{a string of pixel height}

\item{elementId}{a unique id for the htmlwidget object}
}
\value{
the same myIO object
}
\description{
myIO R + d3.js project
}
