% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAxisFormat.R
\name{setAxisFormat}
\alias{setAxisFormat}
\title{Set Axis Format}
\usage{
setAxisFormat(
  myIO,
  xAxis = "s",
  yAxis = "s",
  toolTip = "s",
  xLabel = NULL,
  yLabel = NULL
)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a string indicating a d3.js format}

\item{yAxis}{a string indicating a d3.js format}

\item{toolTip}{a string indicating a d3.js format}

\item{xLabel}{a string label for axis}

\item{yLabel}{a string label for axis}
}
\value{
A modified \code{myIO} htmlwidget object with updated axis format
  configuration. with options set for the tooltip formats
}
\description{
Sets axis for x axis, y axis, and/or tool tip
}
\examples{
# Set axis formats using d3.js format strings
myIO() |> setAxisFormat(xAxis = ".0f", yAxis = ".1f")

# Set axis labels
myIO() |> setAxisFormat(xLabel = "Weight (lbs)", yLabel = "MPG")

}
