% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAxisFormat.R
\name{setAxisFormat}
\alias{setAxisFormat}
\title{setAxisFormat()}
\usage{
setAxisFormat(myIO, xAxis = NULL, yAxis = NULL, toolTip = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xAxis}{a string indicating a d3.js format}

\item{yAxis}{a string indicating a d3.js format}

\item{toolTip}{a string indicating a d3.js format}
}
\value{
the same myIO object with options set for the tooltip formats
}
\description{
Sets axis for x axis, y axis, and/or tool tip
}
