% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAxisLimits.R
\name{setAxisLimits}
\alias{setAxisLimits}
\title{setAxisLimits()}
\usage{
setAxisLimits(myIO, xlim = list(min = NULL, max = NULL), ylim = list(min =
  NULL, max = NULL))
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xlim}{a list of min and max values}

\item{ylim}{a list of min and max values}
}
\value{
the same myIO object
}
\description{
Sets margins for the top, bottom, left, and right sides of the chart
}
