% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColorScheme.R
\name{setColorScheme}
\alias{setColorScheme}
\title{setColorScheme()}
\usage{
setColorScheme(
  myIO,
  colorScheme = NULL,
  colorShceme = NULL,
  setCategories = NULL
)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{colorScheme}{a vector of colors in the order you want them used}

\item{colorShceme}{a vector of colors in the order you want them used}

\item{setCategories}{an optional vector of names that will be mapped to the corresponding color in the colorScheme}
}
\value{
the same myIO object
}
\description{
Sets color scheme for a chart and the category names (optional)
}
\examples{
# Set a custom color scheme
myIO() |> setColorScheme(colorScheme = list("red", "blue", "green"))

# Set colors with category labels
myIO() |> setColorScheme(
  colorScheme = list("steelblue", "orange"),
  setCategories = c("Group A", "Group B")
)

}
