% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColorScheme.R
\name{setColorScheme}
\alias{setColorScheme}
\title{setColorScheme()}
\usage{
setColorScheme(myIO, colorShceme, setCategories = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{colorShceme}{a vector of colors in the order you want them used}

\item{setCategories}{an optional vector of names that will be mapped to the corresponding color in the colorScheme}
}
\value{
the same myIO object
}
\description{
Sets color scheme for a chart and the category names (optional)
}
