% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMargin.R
\name{setMargin}
\alias{setMargin}
\title{setMargin()}
\usage{
setMargin(myIO, top = 20, bottom = 40, left = 50, right = 50)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{top}{a numeric value representing in pixels the top margin}

\item{bottom}{a numeric value representing in pixels the bottom margin}

\item{left}{a numeric value representing in pixels the left margin}

\item{right}{a numeric value representing in pixels the right margin}
}
\value{
the same myIO object
}
\description{
Sets margins for the top, bottom, left, and right sides of the chart
}
\examples{
# Set custom margins
myIO() |> setMargin(top = 50, bottom = 80, left = 60, right = 20)

}
