% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReferenceLines.R
\name{setReferenceLines}
\alias{setReferenceLines}
\title{setReferenceLines()}
\usage{
setReferenceLines(myIO, xRef = 0, yRef = 0)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{xRef}{a list of the reference line value of x}

\item{yRef}{a list of the reference line value of y}
}
\value{
the same myIO object
}
\description{
Sets x and y reference lines
}
