% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setToggle.R
\name{setToggle}
\alias{setToggle}
\title{setToggle()}
\usage{
setToggle(myIO, newY, newScaleY = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{newY}{a string indicating the variable name in data for toggle}
}
\value{
the same myIO object
}
\description{
Sets toggle options for y_var and adds a toggle button for chart
}
