% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setToolTipOptions.R
\name{setToolTipOptions}
\alias{setToolTipOptions}
\title{setToolTipOptions()}
\usage{
setToolTipOptions(myIO, suppressY = NULL)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{suppressY}{a boolean}
}
\value{
the same myIO object with options set for the tooltip formats
}
\description{
Generic function for setting tool tip options for a chart
}
