% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTransitionSpeed.R
\name{setTransitionSpeed}
\alias{setTransitionSpeed}
\title{Set Transition Speed}
\usage{
setTransitionSpeed(myIO, speed)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{speed}{a number indicating the speed of transition in milliseconds}
}
\value{
A modified \code{myIO} htmlwidget object with updated transition
  speed.
}
\description{
Sets transition speeds across the chart (set to 0 to suppress)
}
\examples{
# Set transition speed to 500ms
myIO() |> setTransitionSpeed(speed = 500)

# Disable transitions
myIO() |> setTransitionSpeed(speed = 0)

}
