% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTransitionSpeed.R
\name{setTransitionSpeed}
\alias{setTransitionSpeed}
\title{setTransitionSpeed()}
\usage{
setTransitionSpeed(myIO, speed)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{speed}{a number indicating the speed of transition in milliseconds}
}
\value{
the same myIO object
}
\description{
Sets transition speeds across the chart (set to 0 to suppress)
}
