% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_duckdb_wasm.R
\name{stop_duckdb_wasm_missing}
\alias{stop_duckdb_wasm_missing}
\title{Raise a DuckDB-WASM missing-runtime condition}
\usage{
stop_duckdb_wasm_missing()
}
\description{
Raise a DuckDB-WASM missing-runtime condition
}
\keyword{internal}
