% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressLegend.R
\name{suppressLegend}
\alias{suppressLegend}
\title{suppressLegend()}
\usage{
suppressLegend(myIO, suppressLegend = TRUE)
}
\arguments{
\item{myIO}{an htmlwidget object created by the myIO() function}

\item{suppressLegend}{a logical operator defining whether the legend should be printed or not}
}
\value{
the same myIO object
}
\description{
Suppresses legend from printing
}
